requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.11.9
pydantic-settings==2.11.0
python-dotenv==1.1.1
weaviate-client==4.9.6
anthropic==0.45.1
openai==2.1.0
neo4j==5.28.2
httpx==0.28.1
